Changed decimal to int

+++ DieteticSNS/DieteticSNS.Domain/Entities/Ingredient.cs
@@ -7,9 +7,9 @@ namespace DieteticSNS.Domain.Entities
     public class Ingredient : BaseEntity<int>
     {
         public string Name { get; set; }
-        public decimal Protein { get; set; }
-        public decimal Carbohydrate { get; set; }
-        public decimal Fat { get; set; }
+        public int Protein { get; set; }
+        public int Carbohydrate { get; set; }
+        public int Fat { get; set; }
 
         public virtual ICollection<RecipeIngredient> RecipeIngredients { get; set; } = new HashSet<RecipeIngredient>();
     }
